refactor contract details components
