+++ ui/address/contract/__screenshots__/ContractDetails.pw.tsx_default_with-certified-icon-mobile-1.png
