+++ ui/address/contract/__screenshots__/ContractDetails.pw.tsx_mobile_with-proxy-address-alert-mobile-1.png
