+++ ui/address/contract/alerts/ContractDetailsAlertProxyPattern.tsx
